code now compatible with pep8
