+++ sktime/transformers/series_as_features/dictionary_based/_paa.py
@@ -3,7 +3,6 @@ from sktime.transformers.series_as_features.base import \
     BaseSeriesAsFeaturesTransformer
 from sktime.utils.data_container import tabularize
 from sktime.utils.validation.series_as_features import check_X
-from sktime.utils.load_data import load_from_tsfile_to_dataframe as load_ts
 
 __author__ = "Matthew Middlehurst"
